// src/types/models/rp.ts
export interface Rp {
  _id: string;
  siteBoundId: string;
  slopeAngle: number;
  crepeAngle: number;
  volume: number;
  sizeX: number;
  sizeY: number;
  sizeZ: number;
  positionX: number;
  positionY: number;
  positionZ: number;
  rotationX: number;
  rotationY: number;
  rotationZ: number;
  isSelected?: boolean;
}